Move test_help.hpp to include directory.
