+++ unittest/tp_validation_reader_test.cxx
@@ -10,7 +10,7 @@
 #define FMT_HEADER_ONLY
 
 #include "tpglibs/testapp/tp/TPValidationReader.hpp"
-#include "test_helpers.hpp"
+#include "include/test_helpers.hpp"
 
 #include <boost/test/unit_test.hpp>
 #include <fstream>
